refactor(shell): trocear FolderTabView en folder-tab/ — fase 2 (03/T02)
FolderTabView baja de 1.194 a ~565 líneas extrayendo sin cambios de
comportamiento: folderTabShared (color de carpeta, iconos/etiquetas de
tipo), ColorPickerPopover, SubfolderCard, FileRow, NewFolderInline y
AddMenu, cada uno en su archivo bajo shell/folder-tab/. Imports muertos
podados; API del componente intacta.
