+++ app/components/shell/FolderTabView.tsx
@@ -1,12 +1,9 @@
 import { useMemo, useCallback, useState, useRef, useEffect, Fragment } from 'react';
-import { CONTENT_PINK, FOLDER_TAB_SWATCHES, FOLDER_COLOR_DEFAULT } from '@/lib/ui/palettes';
 import { Modal, ScrollArea, Stack, UnstyledButton, Text, Group } from '@mantine/core';
 import DomeButton from '@/components/ui/DomeButton';
-import { formatDistanceToNow } from 'date-fns';
 import {
-  FolderOpen, Folder, FileText, Plus, Home, ChevronRight,
-  MoreVertical, Trash2, Pencil, X, Check, Upload, Link2, ChevronDown,
-  Palette, FolderInput,
+  FolderOpen, Plus, Home, ChevronRight,
+  Pencil, X, Check, Upload,
 } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import { useShallow } from 'zustand/react/shallow';
@@ -15,7 +12,6 @@ import { useTabStore } from '@/lib/store/useTabStore';
 import { useAppStore } from '@/lib/store/useAppStore';
 import MoveToProjectModal, { filterMoveProjectRoots } from '@/components/workspace/MoveToProjectModal';
 import SelectionActionBar from '@/components/home/SelectionActionBar';
-import DomeResourceIcon from '@/components/ui/DomeResourceIcon';
 
 // ─── Types ────────────────────────────────────────────────────────────────────
 
@@ -26,634 +22,13 @@ interface FolderTabViewProps {
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
 
-function getFolderColor(folder: Resource): string {
-  const meta = folder.metadata as { color?: string } | undefined;
-  return meta?.color ?? 'var(--dome-text-muted)';
-}
-
-function ResourceTypeIcon({ type, name, className }: { type: string; name?: string; className?: string }) {
-  return (
-    <DomeResourceIcon
-      type={type}
-      name={name}
-      size={16}
-      className={className ?? 'size-4 shrink-0'}
-      strokeWidth={1.75}
-    />
-  );
-}
-
-const TYPE_LABELS: Record<string, string> = {
-  note: 'Nota', notebook: 'Cuaderno', url: 'URL',
-  pdf: 'PDF', image: 'Imagen', video: 'Video',
-  audio: 'Audio', document: 'Documento', ppt: 'Presentación',
-};
-
-const TYPE_COLORS: Record<string, string> = {
-  note: 'var(--accent)', notebook: 'var(--accent)', url: 'var(--success)',
-  pdf: 'var(--error)', image: 'var(--warning)', video: CONTENT_PINK, audio: 'var(--accent)', ppt: 'var(--warning)',
-};
-
-// ─── ColorPickerPopover ───────────────────────────────────────────────────────
-
-const SWATCHES = FOLDER_TAB_SWATCHES;
-
-function ColorPickerPopover({
-  pos, currentColor, onSave, onClose,
-}: {
-  pos: { top: number; left: number };
-  currentColor: string;
-  onSave: (color: string) => void;
-  onClose: () => void;
-}) {
-  const { t } = useTranslation();
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const close = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) onClose();
-    };
-    document.addEventListener('mousedown', close);
-    return () => document.removeEventListener('mousedown', close);
-  }, [onClose]);
-
-  return (
-    // onMouseDown only stops the outside-click closer (propagation barrier).
-    // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-    <div
-      ref={ref}
-      role="group"
-      aria-label={t('folder.changeColor', 'Cambiar color')}
-      className="fixed z-[var(--z-popover)] rounded-xl shadow-lg p-2.5"
-      style={{ top: pos.top, left: pos.left, background: 'var(--dome-surface)', border: '1px solid var(--dome-border)' }}
-      onMouseDown={(e) => e.stopPropagation()}
-    >
-      <div className="grid grid-cols-6 gap-1.5">
-        {SWATCHES.map((color) => (
-          <button
-            key={color}
-            type="button"
-            aria-label={color}
-            onClick={() => { onSave(color); onClose(); }}
-            className="size-6 rounded-md transition-all hover:scale-110"
-            style={{
-              backgroundColor: color,
-              border: currentColor.toLowerCase() === color.toLowerCase()
-                ? '2px solid var(--dome-accent)'
-                : '2px solid transparent',
-              outline: currentColor.toLowerCase() === color.toLowerCase()
-                ? '1px solid var(--dome-accent)'
-                : 'none',
-              outlineOffset: 1,
-              cursor: 'pointer',
-            }}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// ─── SubfolderCard ────────────────────────────────────────────────────────────
-
-function SubfolderCard({
-  folder,
-  onClick,
-  onRename,
-  onDelete,
-  onChangeColor,
-  onMoveToProject,
-  selected,
-  showSelectionChrome,
-  onToggleSelect,
-}: {
-  folder: Resource;
-  onClick: () => void;
-  onRename: (newTitle: string) => void;
-  onDelete: () => void;
-  onChangeColor: (color: string) => void;
-  onMoveToProject: () => void;
-  selected: boolean;
-  showSelectionChrome: boolean;
-  onToggleSelect: (e: React.MouseEvent) => void;
-}) {
-  const { t } = useTranslation();
-  const [hovered, setHovered] = useState(false);
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [menuPos, setMenuPos] = useState<{ top: number; right: number } | null>(null);
-  const [renaming, setRenaming] = useState(false);
-  const [renameValue, setRenameValue] = useState(folder.title ?? '');
-  const [colorPickerPos, setColorPickerPos] = useState<{ top: number; left: number } | null>(null);
-  const menuBtnRef = useRef<HTMLButtonElement>(null);
-  const renameRef = useRef<HTMLInputElement>(null);
-  const color = getFolderColor(folder);
-  const folderSelectId = `folder-tab-row-select-${folder.id}`;
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    const close = () => setMenuOpen(false);
-    document.addEventListener('mousedown', close);
-    return () => document.removeEventListener('mousedown', close);
-  }, [menuOpen]);
-
-  useEffect(() => {
-    if (renaming) renameRef.current?.focus();
-  }, [renaming]);
-
-  const commitRename = () => {
-    if (renameValue.trim() && renameValue.trim() !== folder.title) {
-      onRename(renameValue.trim());
-    }
-    setRenaming(false);
-  };
-
-  const openMenu = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (!menuOpen && menuBtnRef.current) {
-      const rect = menuBtnRef.current.getBoundingClientRect();
-      setMenuPos({ top: rect.bottom + 4, right: window.innerWidth - rect.right });
-    }
-    setMenuOpen((v) => !v);
-  };
-
-  const menuItem = (icon: React.ReactNode, label: string, action: () => void, danger = false) => (
-    <button
-      type="button"
-      onClick={(e) => { e.stopPropagation(); setMenuOpen(false); action(); }}
-      className="flex items-center gap-2 w-full px-3 py-2 text-xs text-left"
-      style={{ color: danger ? 'var(--dome-error)' : 'var(--dome-text)', background: 'none', border: 'none', cursor: 'pointer' }}
-      onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-      onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-    >
-      {icon} {label}
-    </button>
-  );
-
-  return (
-    <div
-      className="relative flex flex-col rounded-xl transition-all"
-      style={{
-        background: hovered ? 'var(--dome-bg-hover)' : 'var(--dome-surface)',
-        border: `1px solid ${selected ? 'var(--dome-accent)' : hovered ? color : 'var(--dome-border)'}`,
-      }}
-      onMouseEnter={() => setHovered(true)}
-      onMouseLeave={() => { setHovered(false); }}
-    >
-      {showSelectionChrome ? (
-        // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions -- stopPropagation barrier so the card click doesn't fire
-        <label
-          className="absolute left-2 top-2 z-[2] flex items-center cursor-pointer"
-          htmlFor={folderSelectId}
-          onMouseDown={(e) => e.stopPropagation()}
-        >
-          <input
-            id={folderSelectId}
-            type="checkbox"
-            checked={selected}
-            onChange={() => {}}
-            onClick={(e) => { e.stopPropagation(); onToggleSelect(e); }}
-            className="rounded border cursor-pointer"
-            style={{ accentColor: 'var(--dome-accent)' }}
-            aria-label={t('selection.deselect')}
-          />
-        </label>
-      ) : null}
-      {/* Main clickable area */}
-      {renaming ? (
-        <div className="flex items-center gap-2 px-3 py-2.5">
-          <Folder className="size-4 shrink-0" style={{ color }} />
-          <input
-            ref={renameRef}
-            type="text"
-            value={renameValue}
-            onChange={(e) => setRenameValue(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') commitRename();
-              if (e.key === 'Escape') { setRenaming(false); setRenameValue(folder.title ?? ''); }
-            }}
-            onClick={(e) => e.stopPropagation()}
-            className="flex-1 min-w-0 text-sm outline-none rounded px-1.5 py-0.5"
-            style={{ background: 'var(--dome-bg)', border: '1px solid var(--dome-accent)', color: 'var(--dome-text)' }}
-          />
-          <button type="button" onClick={(e) => { e.stopPropagation(); commitRename(); }}
-            style={{ color: 'var(--dome-accent)', background: 'none', border: 'none', cursor: 'pointer' }}>
-            <Check className="size-3.5" />
-          </button>
-          <button type="button" onClick={(e) => { e.stopPropagation(); setRenaming(false); setRenameValue(folder.title ?? ''); }}
-            style={{ color: 'var(--dome-text-muted)', background: 'none', border: 'none', cursor: 'pointer' }}>
-            <X className="size-3.5" />
-          </button>
-        </div>
-      ) : (
-        <button
-          type="button"
-          onClick={(e) => {
-            if (e.metaKey || e.ctrlKey) {
-              e.preventDefault();
-              onToggleSelect(e);
-              return;
-            }
-            onClick();
-          }}
-          className="flex items-center gap-2.5 px-3 py-2.5 text-left w-full"
-          style={{
-            background: 'none',
-            border: 'none',
-            cursor: 'pointer',
-            paddingLeft: showSelectionChrome ? 36 : undefined,
-          }}
-        >
-          <Folder className="size-4 shrink-0" style={{ color }} />
-          <span className="text-sm font-medium truncate flex-1 min-w-0" style={{ color: 'var(--dome-text)' }}>
-            {folder.title}
-          </span>
-        </button>
-      )}
-
-      {/* 3-dot menu button */}
-      {!renaming && (hovered || menuOpen) && (
-        <div className="absolute top-1.5 right-1.5">
-          <button
-            ref={menuBtnRef}
-            type="button"
-            onClick={openMenu}
-            className="flex items-center justify-center size-5 rounded-md transition-colors"
-            style={{ color: 'var(--dome-text-muted)', background: menuOpen ? 'var(--dome-bg-hover)' : 'none', border: 'none', cursor: 'pointer' }}
-            onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-            onMouseLeave={(e) => { if (!menuOpen) (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-          >
-            <MoreVertical className="size-3" />
-          </button>
-
-          {menuOpen && menuPos && (
-            <div
-              role="menu"
-              tabIndex={-1}
-              className="fixed z-[var(--z-popover)] rounded-lg shadow-lg py-1 min-w-[150px]"
-              style={{ background: 'var(--dome-surface)', border: '1px solid var(--dome-border)', top: menuPos.top, right: menuPos.right }}
-              onMouseDown={(e) => e.stopPropagation()}
-            >
-              {menuItem(<Pencil className="size-3" />, t('folder.rename'), () => { setRenaming(true); setRenameValue(folder.title ?? ''); })}
-              {menuItem(<Palette className="size-3" />, t('folder.changeColor', 'Cambiar color'), () => {
-                if (menuBtnRef.current) {
-                  const rect = menuBtnRef.current.getBoundingClientRect();
-                  setColorPickerPos({ top: rect.bottom + 4, left: Math.max(4, rect.right - 220) });
-                }
-              })}
-              {menuItem(<FolderInput className="size-3" />, t('selection.move_to_project'), onMoveToProject)}
-              <div style={{ height: 1, background: 'var(--dome-border)', margin: '3px 0' }} />
-              {menuItem(<Trash2 className="size-3" />, t('folder.delete'), onDelete, true)}
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Color picker popover */}
-      {colorPickerPos && (
-        <ColorPickerPopover
-          pos={colorPickerPos}
-          currentColor={color.startsWith('#') ? color : FOLDER_COLOR_DEFAULT}
-          onSave={onChangeColor}
-          onClose={() => setColorPickerPos(null)}
-        />
-      )}
-    </div>
-  );
-}
-
-// ─── FileRow ──────────────────────────────────────────────────────────────────
-
-function FileRow({
-  file,
-  isLast,
-  onOpen,
-  onDelete,
-  onRename,
-  onMoveToProject,
-  selected,
-  showSelectionChrome,
-  onToggleSelect,
-}: {
-  file: Resource;
-  isLast: boolean;
-  onOpen: () => void;
-  onDelete: () => void;
-  onRename: (newTitle: string) => void;
-  onMoveToProject: () => void;
-  selected: boolean;
-  showSelectionChrome: boolean;
-  onToggleSelect: (e: React.MouseEvent) => void;
-}) {
-  const { t } = useTranslation();
-  const [hovered, setHovered] = useState(false);
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [menuPos, setMenuPos] = useState<{ top: number; right: number } | null>(null);
-  const [renaming, setRenaming] = useState(false);
-  const [renameValue, setRenameValue] = useState(file.title ?? '');
-  const menuBtnRef = useRef<HTMLButtonElement>(null);
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    const close = () => setMenuOpen(false);
-    document.addEventListener('mousedown', close);
-    return () => document.removeEventListener('mousedown', close);
-  }, [menuOpen]);
-
-  const typeColor = TYPE_COLORS[file.type] ?? 'var(--dome-text-muted)';
-  const typeLabel = TYPE_LABELS[file.type] ?? file.type;
-  const timeAgo = file.updated_at
-    ? formatDistanceToNow(new Date(file.updated_at), { addSuffix: true })
-    : null;
-
-  const commitRename = () => {
-    if (renameValue.trim() && renameValue.trim() !== file.title) {
-      onRename(renameValue.trim());
-    }
-    setRenaming(false);
-  };
-
-  return (
-    <div
-      className="flex items-center gap-3 px-4 py-2.5 transition-colors relative"
-      style={{
-        borderBottom: isLast ? undefined : '1px solid var(--dome-border)',
-        background: hovered ? 'var(--dome-bg-hover)' : 'var(--dome-surface)',
-        outline: selected ? '1px solid var(--dome-accent)' : undefined,
-        outlineOffset: -1,
-      }}
-      onMouseEnter={() => setHovered(true)}
-      onMouseLeave={() => { setHovered(false); setMenuOpen(false); }}
-    >
-      {showSelectionChrome ? (
-        <input
-          type="checkbox"
-          checked={selected}
-          onChange={() => {}}
-          onClick={(e) => { e.stopPropagation(); onToggleSelect(e); }}
-          className="rounded border shrink-0 cursor-pointer"
-          style={{ accentColor: 'var(--dome-accent)' }}
-          aria-label={t('selection.deselect')}
-        />
-      ) : null}
-      <div className="size-1.5 rounded-full shrink-0" style={{ background: typeColor }} />
-      <div style={{ color: typeColor }}>
-        <ResourceTypeIcon type={file.type} name={file.title} />
-      </div>
-
-      {renaming ? (
-        <div className="flex-1 flex items-center gap-2 min-w-0">
-          <input
-            type="text"
-            value={renameValue}
-            onChange={(e) => setRenameValue(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') commitRename();
-              if (e.key === 'Escape') setRenaming(false);
-            }}
-            className="flex-1 text-[13px] font-medium rounded px-2 py-0.5 outline-none"
-            style={{ background: 'var(--dome-bg)', border: '1px solid var(--dome-accent)', color: 'var(--dome-text)' }}
-          />
-          <button type="button" onClick={commitRename} style={{ color: 'var(--dome-accent)', background: 'none', border: 'none', cursor: 'pointer' }}>
-            <Check className="size-3.5" />
-          </button>
-          <button type="button" onClick={() => setRenaming(false)} style={{ color: 'var(--dome-text-muted)', background: 'none', border: 'none', cursor: 'pointer' }}>
-            <X className="size-3.5" />
-          </button>
-        </div>
-      ) : (
-        <button
-          type="button"
-          onClick={(e) => {
-            if (e.metaKey || e.ctrlKey) {
-              e.preventDefault();
-              onToggleSelect(e);
-              return;
-            }
-            onOpen();
-          }}
-          className="flex-1 text-left text-[13px] font-medium truncate hover:underline underline-offset-2 min-w-0"
-          style={{ color: 'var(--dome-text)', background: 'none', border: 'none', cursor: 'pointer', padding: 0 }}
-        >
-          {file.title || t('folder.untitled')}
-        </button>
-      )}
-
-      <span
-        className="text-[10px] px-1.5 py-0.5 rounded-md shrink-0 font-medium"
-        style={{ background: `${typeColor}18`, color: typeColor }}
-      >
-        {typeLabel}
-      </span>
-
-      {timeAgo && (
-        <span className="text-[11px] shrink-0 tabular-nums" style={{ color: 'var(--dome-text-muted)' }}>
-          {timeAgo}
-        </span>
-      )}
-
-      {hovered && !renaming && (
-        <div className="shrink-0">
-          <button
-            ref={menuBtnRef}
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              if (!menuOpen && menuBtnRef.current) {
-                const rect = menuBtnRef.current.getBoundingClientRect();
-                setMenuPos({ top: rect.bottom + 4, right: window.innerWidth - rect.right });
-              }
-              setMenuOpen((v) => !v);
-            }}
-            className="flex items-center justify-center rounded p-0.5"
-            style={{ color: 'var(--dome-text-muted)', background: 'none', border: 'none', cursor: 'pointer' }}
-          >
-            <MoreVertical className="size-3.5" />
-          </button>
-          {menuOpen && menuPos && (
-            <div
-              role="menu"
-              tabIndex={-1}
-              className="fixed z-[var(--z-popover)] rounded-lg shadow-lg py-1 min-w-[130px]"
-              style={{ background: 'var(--dome-surface)', border: '1px solid var(--dome-border)', top: menuPos.top, right: menuPos.right }}
-              onMouseDown={(e) => e.stopPropagation()}
-            >
-              <button
-                type="button"
-                onClick={() => { setMenuOpen(false); setRenaming(true); setRenameValue(file.title ?? ''); }}
-                className="flex items-center gap-2 w-full px-3 py-2 text-xs text-left"
-                style={{ color: 'var(--dome-text)', background: 'none', border: 'none', cursor: 'pointer' }}
-                onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-                onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-              >
-                <Pencil className="size-3" /> {t('folder.rename')}
-              </button>
-              <button
-                type="button"
-                onClick={() => { setMenuOpen(false); onMoveToProject(); }}
-                className="flex items-center gap-2 w-full px-3 py-2 text-xs text-left"
-                style={{ color: 'var(--dome-text)', background: 'none', border: 'none', cursor: 'pointer' }}
-                onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-                onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-              >
-                <FolderInput className="size-3" /> {t('selection.move_to_project')}
-              </button>
-              <button
-                type="button"
-                onClick={() => { setMenuOpen(false); onDelete(); }}
-                className="flex items-center gap-2 w-full px-3 py-2 text-xs text-left"
-                style={{ color: 'var(--dome-error)', background: 'none', border: 'none', cursor: 'pointer' }}
-                onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-                onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-              >
-                <Trash2 className="size-3" /> {t('folder.delete')}
-              </button>
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── NewFolderInline ──────────────────────────────────────────────────────────
-
-function NewFolderInline({ onConfirm, onCancel }: { onConfirm: (name: string) => void; onCancel: () => void }) {
-  const { t } = useTranslation();
-  const [value, setValue] = useState('');
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => { inputRef.current?.focus(); }, []);
-
-  const handleConfirm = () => { if (value.trim()) onConfirm(value.trim()); };
-
-  return (
-    <div
-      className="flex flex-col w-full rounded-xl overflow-hidden"
-      style={{ border: '1.5px dashed var(--dome-border)', background: 'var(--dome-surface)' }}
-    >
-      <div className="flex items-center gap-2 px-3 py-2.5 min-w-0">
-        <Folder className="size-4 shrink-0" style={{ color: 'var(--dome-text-muted)' }} />
-        <input
-          ref={inputRef}
-          type="text"
-          value={value}
-          onChange={(e) => setValue(e.target.value)}
-          placeholder={t('folder.folderNamePlaceholder')}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter') handleConfirm();
-            if (e.key === 'Escape') onCancel();
-          }}
-          className="text-sm outline-none bg-transparent flex-1 min-w-0 truncate"
-          style={{ color: 'var(--dome-text)', border: 'none', padding: 0 }}
-        />
-      </div>
-      <div
-        className="flex items-center justify-end gap-1 px-2 py-1.5"
-        style={{ borderTop: '1px solid var(--dome-border)' }}
-      >
-        <button
-          type="button"
-          onClick={handleConfirm}
-          disabled={!value.trim()}
-          className="flex items-center justify-center size-6 rounded-md transition-colors disabled:opacity-40"
-          style={{ color: 'var(--dome-accent)', background: 'none', border: 'none', cursor: value.trim() ? 'pointer' : 'default' }}
-          onMouseEnter={(e) => { if (value.trim()) (e.currentTarget as HTMLButtonElement).style.background = 'rgba(124,111,205,0.1)'; }}
-          onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-        >
-          <Check className="size-3.5" />
-        </button>
-        <button
-          type="button"
-          onClick={onCancel}
-          className="flex items-center justify-center size-6 rounded-md transition-colors"
-          style={{ color: 'var(--dome-text-muted)', background: 'none', border: 'none', cursor: 'pointer' }}
-          onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-          onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-        >
-          <X className="size-3.5" />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// ─── AddMenu ─────────────────────────────────────────────────────────────────
-
-function AddMenu({ onNewNote, onNewFolder, onUpload, onAddUrl }: {
-  onNewNote: () => void;
-  onNewFolder: () => void;
-  onUpload: () => void;
-  onAddUrl: () => void;
-}) {
-  const { t } = useTranslation();
-  const [open, setOpen] = useState(false);
-  const btnRef = useRef<HTMLButtonElement>(null);
-  const menuRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) return;
-    const close = (e: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node) &&
-          btnRef.current && !btnRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', close);
-    return () => document.removeEventListener('mousedown', close);
-  }, [open]);
-
-  const item = (icon: React.ReactNode, label: string, onClick: () => void, color?: string) => (
-    <button
-      type="button"
-      onClick={() => { setOpen(false); onClick(); }}
-      className="flex items-center gap-2.5 w-full px-3 py-2.5 text-sm text-left transition-colors"
-      style={{ color: color ?? 'var(--dome-text)', background: 'none', border: 'none', cursor: 'pointer' }}
-      onMouseEnter={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'var(--dome-bg-hover)'; }}
-      onMouseLeave={(e) => { (e.currentTarget as HTMLButtonElement).style.background = 'none'; }}
-    >
-      {icon}
-      {label}
-    </button>
-  );
-
-  return (
-    <div className="relative">
-      <button
-        ref={btnRef}
-        type="button"
-        onClick={() => setOpen((v) => !v)}
-        className="flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-sm font-medium transition-all"
-        style={{
-          background: 'var(--dome-accent)',
-          color: 'var(--dome-on-accent)',
-          border: 'none',
-          cursor: 'pointer',
-          boxShadow: '0 2px 8px rgba(124,111,205,0.35)',
-        }}
-      >
-        <Plus className="size-3.5" />
-        {t('folder.addBtn', 'Añadir')}
-        <ChevronDown className={`size-3 transition-transform ${open ? 'rotate-180' : ''}`} />
-      </button>
-
-      {open && (
-        <div
-          ref={menuRef}
-          className="absolute right-0 top-full mt-1.5 z-[var(--z-popover)] rounded-xl shadow-xl py-1.5 min-w-[200px]"
-          style={{ background: 'var(--dome-surface)', border: '1px solid var(--dome-border)' }}
-        >
-          {item(<FileText className="size-4" style={{ color: 'var(--dome-accent)' }} />, t('toolbar.note', 'Nueva nota'), onNewNote)}
-          {item(<Upload className="size-4" style={{ color: 'var(--accent)' }} />, t('toolbar.import', 'Subir archivo'), onUpload)}
-          {item(<Link2 className="size-4" style={{ color: 'var(--success)' }} />, t('toolbar.link', 'Añadir enlace'), onAddUrl)}
-          <div className="my-1" style={{ height: 1, background: 'var(--dome-border)' }} />
-          {item(<Folder className="size-4" style={{ color: 'var(--dome-text-muted)' }} />, t('folder.newFolderBtn', 'Nueva carpeta'), onNewFolder)}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── FolderTabView ────────────────────────────────────────────────────────────
+// Subcomponentes extraídos (03/T02) — misma UI, archivos por pieza.
+import { getFolderColor } from './folder-tab/folderTabShared';
+import ColorPickerPopover from './folder-tab/ColorPickerPopover';
+import SubfolderCard from './folder-tab/SubfolderCard';
+import FileRow from './folder-tab/FileRow';
+import NewFolderInline from './folder-tab/NewFolderInline';
+import AddMenu from './folder-tab/AddMenu';
 
 export default function FolderTabView({ folderId, folderTitle }: FolderTabViewProps) {
   const { t } = useTranslation();
