+++ app/components/shell/folder-tab/ColorPickerPopover.tsx
@@ -1,0 +1,66 @@
+/** Color swatch popover for folder cards (03/T02 — extracted from FolderTabView.tsx). */
+
+import { useEffect, useRef } from 'react';
+import { useTranslation } from 'react-i18next';
+import { FOLDER_TAB_SWATCHES } from '@/lib/ui/palettes';
+
+const SWATCHES = FOLDER_TAB_SWATCHES;
+
+export default function ColorPickerPopover({
+  pos, currentColor, onSave, onClose,
+}: {
+  pos: { top: number; left: number };
+  currentColor: string;
+  onSave: (color: string) => void;
+  onClose: () => void;
+}) {
+  const { t } = useTranslation();
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const close = (e: MouseEvent) => {
+      if (ref.current && !ref.current.contains(e.target as Node)) onClose();
+    };
+    document.addEventListener('mousedown', close);
+    return () => document.removeEventListener('mousedown', close);
+  }, [onClose]);
+
+  return (
+    // onMouseDown only stops the outside-click closer (propagation barrier).
+    // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
+    <div
+      ref={ref}
+      role="group"
+      aria-label={t('folder.changeColor', 'Cambiar color')}
+      className="fixed z-[var(--z-popover)] rounded-xl shadow-lg p-2.5"
+      style={{ top: pos.top, left: pos.left, background: 'var(--dome-surface)', border: '1px solid var(--dome-border)' }}
+      onMouseDown={(e) => e.stopPropagation()}
+    >
+      <div className="grid grid-cols-6 gap-1.5">
+        {SWATCHES.map((color) => (
+          <button
+            key={color}
+            type="button"
+            aria-label={color}
+            onClick={() => { onSave(color); onClose(); }}
+            className="size-6 rounded-md transition-all hover:scale-110"
+            style={{
+              backgroundColor: color,
+              border: currentColor.toLowerCase() === color.toLowerCase()
+                ? '2px solid var(--dome-accent)'
+                : '2px solid transparent',
+              outline: currentColor.toLowerCase() === color.toLowerCase()
+                ? '1px solid var(--dome-accent)'
+                : 'none',
+              outlineOffset: 1,
+              cursor: 'pointer',
+            }}
+          />
+        ))}
+      </div>
+    </div>
+  );
+}
+
+// ─── SubfolderCard ────────────────────────────────────────────────────────────
+
